fixed msvc error about use of size_t in "omp paraller for"

+++ mandelbulber2/src/voxel_export.hpp
@@ -64,7 +64,7 @@ private:
 	bool StoreLayer(int z) const;
 
 	QScopedArrayPointer<unsigned char> voxelLayer;
-	size_t w, h, l;
+	long long w, h, l;
 	CVector3 limitMin;
 	CVector3 limitMax;
 	QDir folder;
